requirements.txt
fastapi
uvicorn
pydantic
lightrag-hku
python-multipart
psycopg2-binary
python-dotenv
asyncio
